Fix macOS target
