+++ Sources/Mac/Classes/CarouselSpot.swift
@@ -191,26 +191,3 @@ public class CarouselSpot: NSObject, Gridable {
     collectionView.frame.size.height = scrollView.frame.size.height + titleView.frame.size.height
   }
 }
-
-extension CarouselSpot {
-
-  public func sizeForItemAt(indexPath: NSIndexPath) -> CGSize {
-    var width = component.span > 0
-      ? collectionView.frame.width / CGFloat(component.span)
-      : collectionView.frame.width
-
-    if let layout = layout as? NSCollectionViewFlowLayout {
-      width -= layout.sectionInset.left - layout.sectionInset.right
-      width -= layout.minimumInteritemSpacing
-      width -= layout.minimumLineSpacing
-    }
-
-    if component.items[indexPath.item].size.width == 0.0 {
-      component.items[indexPath.item].size.width = width
-    }
-
-    return CGSize(
-      width: ceil(component.items[indexPath.item].size.width),
-      height: ceil(component.items[indexPath.item].size.height))
-  }
-}
