+++ Sources/Mac/Classes/CollectionViewLeftLayout.swift
