+++ Sources/Mac/Classes/GridSpot.swift
