+++ Sources/Mac/Classes/GridSpotItem.swift
