+++ Sources/Mac/Classes/ListSpot.swift
