+++ Sources/Mac/Classes/ListSpotItem.swift
