+++ Sources/Mac/Classes/SpotsContentView.swift
