+++ Sources/Mac/Classes/SpotsController.swift
