+++ Sources/Mac/Classes/SpotsScrollView.swift
