+++ Sources/Mac/Extensions/CollectionAdapter+NSCollectionViewDelegateFlowLayout.swift
@@ -1,0 +1,8 @@
+import Cocoa
+
+extension CollectionAdapter: NSCollectionViewDelegateFlowLayout {
+
+  public func collectionView(collectionView: NSCollectionView, layout collectionViewLayout: NSCollectionViewLayout, sizeForItemAtIndexPath indexPath: NSIndexPath) -> NSSize {
+    return spot.sizeForItemAt(indexPath)
+  }
+}
