+++ Sources/Mac/Structs/GridRegistry.swift
